Another big refactor - Introducing directory splits
Also made the swipe listener work for the blur overlay, unanimated
