+++ owldroid/src/main/java/com/hedvig/android/owldroid/di/FragmentContributorModule.kt
@@ -1,5 +1,7 @@
-package com.hedvig.android.owldroid
+package com.hedvig.android.owldroid.di
 
+import com.hedvig.android.owldroid.ui.marketing.MarketingFragment
+import com.hedvig.android.owldroid.ui.marketing.StoryFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
