+++ owldroid/src/main/java/com/hedvig/android/owldroid/di/ViewModelKey.kt
@@ -1,0 +1,10 @@
+package com.hedvig.android.owldroid.di
+
+import android.arch.lifecycle.ViewModel
+import dagger.MapKey
+import kotlin.reflect.KClass
+
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
+@Retention(AnnotationRetention.RUNTIME)
+@MapKey
+annotation class ViewModelKey(val value: KClass<out ViewModel>)
