+++ owldroid/src/main/java/com/hedvig/android/owldroid/util/ListExtensions.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.owldroid
+package com.hedvig.android.owldroid.util
 
 val<T> List<T>.tail: List<T>
     get() = subList(1, size)
